requirements.txt
keras
flask
opencv-python
numpy
time
sys
werkzeug.utils
gevent.pywsgi
glob
re
random
matplotlib